Gamps

+++ cpp_project/include/coders/constant/apca/apca_window.h
@@ -3,13 +3,14 @@
 #define CPP_PROJECT_APCA_WINDOW_H
 
 #include <string>
+#include "constants.h"
 #include "pca_window.h"
 
 class APCAWindow: public Window {
 
 private:
-    int min;
-    int max;
+    double min;
+    double max;
 #if !MASK_MODE
     bool nan_window;
 #endif
@@ -20,7 +21,7 @@ public:
     bool isFull();
     bool isEmpty();
     void addFirstValue(std::string x);
-    bool updateConstantValue(int new_min, int new_max);
+    bool updateConstantValue(double new_min, double new_max);
 };
 
 #endif //CPP_PROJECT_APCA_WINDOW_H
